refactor(rats-tls): consolidate shared rustls config into utils/rustls/
Move duplicated TlsConfigGenerator logic, cert verifier wrappers,
and dummy cert infrastructure from ingress/egress into a unified
utils/rustls/ directory structure:

- utils/rustls/config/ — TlsConfigGenerator with Alpn enum,
  LazyOnetimeTlsClientConfig and LazyOnetimeTlsServerConfig with
  handshake_with_stream() that bundles TLS handshake + verity_pending_cert
- utils/rustls/ra/ — LazyCertVerifier (common), LazyServerCertVerifier,
  LazyClientCertVerifier (moved from tunnel/cert_verifier.rs and
  protocol-specific cert_verifier modules)
- utils/rustls/dummy/ — DummyServerCertVerifier, cert PEM files
  (moved from utils/certs/ and ingress cert_verifier/dummy.rs)

Callers now use get_lazy_one_time_rustls_{client,server}_config(Alpn)
instead of separate config files per protocol direction. The verifier
field is private — the only way to get a stream is through
handshake_with_stream(), which always calls verity_pending_cert when
a verifier is configured.
