+++ tng/src/tunnel/mod.rs
@@ -1,7 +1,5 @@
 pub(crate) mod access_log;
 pub(crate) mod attestation_result;
-#[cfg(unix)]
-pub(crate) mod cert_verifier;
 #[cfg(feature = "__egress-common")]
 pub(crate) mod egress;
 pub mod endpoint;
