+++ tng/src/tunnel/utils/rustls/dummy/mod.rs
@@ -1,4 +1,7 @@
-#[allow(dead_code)]
+pub mod verifier;
+
+pub use verifier::DummyServerCertVerifier;
+
 pub const TNG_DUMMY_CERT: &str = include_str!("servercert.pem");
 #[allow(dead_code)]
 pub const TNG_DUMMY_KEY: &str = include_str!("serverkey.pem");
