+++ tng/src/tunnel/utils/rustls/mod.rs
@@ -1,2 +1,3 @@
+pub mod config;
 pub mod dummy;
 pub mod ra;
