+++ tng/src/tunnel/utils/rustls/ra/client_cert_verifier.rs
@@ -11,17 +11,18 @@ use rustls::{
 use tokio_rustls::rustls::RootCertStore;
 
 use crate::tunnel::{
-    attestation_result::AttestationResult, cert_verifier::TngCommonCertVerifier,
-    ra_context::VerifyContext, utils::certs::TNG_DUMMY_CERT,
+    attestation_result::AttestationResult,
+    ra_context::VerifyContext,
+    utils::rustls::{dummy::TNG_DUMMY_CERT, ra::common::LazyCertVerifier},
 };
 
 #[derive(Debug)]
-pub struct TngClientCertVerifier {
+pub struct LazyClientCertVerifier {
     inner: Arc<dyn ClientCertVerifier>,
-    common: TngCommonCertVerifier,
+    common: LazyCertVerifier,
 }
 
-impl TngClientCertVerifier {
+impl LazyClientCertVerifier {
     pub fn new(verify_ctx: Arc<VerifyContext>) -> Result<Self> {
         let mut cert = TNG_DUMMY_CERT.as_bytes();
         let certs = rustls_pemfile::certs(&mut cert).collect::<Result<Vec<_>, _>>()?;
@@ -32,16 +33,16 @@ impl TngClientCertVerifier {
 
         Ok(Self {
             inner: verifier,
-            common: TngCommonCertVerifier::new(verify_ctx),
+            common: LazyCertVerifier::new(verify_ctx),
         })
     }
 
     pub async fn verity_pending_cert(&self) -> Result<AttestationResult> {
-        self.common.verity_pending_cert().await
+        self.common.verify_pending_cert().await
     }
 }
 
-impl rustls::server::danger::ClientCertVerifier for TngClientCertVerifier {
+impl rustls::server::danger::ClientCertVerifier for LazyClientCertVerifier {
     fn root_hint_subjects(&self) -> &[rustls::DistinguishedName] {
         &[]
     }
@@ -53,7 +54,7 @@ impl rustls::server::danger::ClientCertVerifier for TngClientCertVerifier {
         _now: rustls::pki_types::UnixTime,
     ) -> std::result::Result<rustls::server::danger::ClientCertVerified, Error> {
         self.common
-            .verify_cert(end_entity)
+            .set_to_pending_cert(end_entity)
             .map(|()| ClientCertVerified::assertion())
     }
